Add Crud Role

+++ src/Controller/ProduitController.php
@@ -16,9 +16,14 @@ use Symfony\Component\String\Slugger\SluggerInterface;
 #[Route('/produit')]
 class ProduitController extends AbstractController
 {
+    
     #[Route('/', name: 'produit_index', methods: ['GET'])]
     public function index(ProduitRepository $produitRepository): Response
     {
+        if($this->getUser()){
+            return $this->redirectToRoute('presentationProduit');
+        }
+        
         return $this->render('produit/index.html.twig', [
             'produits' => $produitRepository->findAll(),
         ]);
@@ -27,6 +32,10 @@ class ProduitController extends AbstractController
     #[Route('/new', name: 'produit_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager, SluggerInterface $slugger): Response
     {
+        if($this->getUser()){
+            return $this->redirectToRoute('presentationProduit');
+        }
+
         $produit = new Produit();
         $form = $this->createForm(ProduitType::class, $produit);
         $form->handleRequest($request);
@@ -66,6 +75,10 @@ class ProduitController extends AbstractController
     #[Route('/{id}', name: 'produit_show', methods: ['GET'])]
     public function show(Produit $produit): Response
     {
+        if($this->getUser()){
+            return $this->redirectToRoute('presentationProduit');
+        }
+
         return $this->render('produit/show.html.twig', [
             'produit' => $produit,
         ]);
@@ -74,6 +87,10 @@ class ProduitController extends AbstractController
     #[Route('/{id}/edit', name: 'produit_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Produit $produit, EntityManagerInterface $entityManager, SluggerInterface $slugger): Response
     {
+        if($this->getUser()){
+            return $this->redirectToRoute('presentationProduit');
+        }
+
         $form = $this->createForm(ProduitType::class, $produit);
         $form->handleRequest($request);
 
@@ -111,6 +128,10 @@ class ProduitController extends AbstractController
     #[Route('/{id}', name: 'produit_delete', methods: ['POST'])]
     public function delete(Request $request, Produit $produit, EntityManagerInterface $entityManager): Response
     {
+        if($this->getUser()){
+            return $this->redirectToRoute('presentationProduit');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$produit->getId(), $request->request->get('_token'))) {
             $entityManager->remove($produit);
             $entityManager->flush();
